List all pybind11 exceptions and refer to code.
I wanted to know excacly what existed, because I wanted to raise a
`RuntimeError`. In practice, it took me quite some time to find where
(a) the pybind11 exceptions are defined, (b) how the C++ exceptions
are converted to Python ones.

At the end, I discovered everything else is converted to RuntimeError,
so I do not need to do anything.

The link is a result to github search, which may be more robust to
moving code around.