Add sequence constructors taking ssize_t parameter.
This patch makes `ssize_t s = 3; py::bytes{"foo", s};` (and likewise for
str, tuple, and list) compile without a -Wnarrowing (ssize_t to size_t)
warning.  Allowing signed sizes is consistent with the CPython C-API
(PyBytes_FromStringAndSize) which also takes a Py_ssize_t parameter, and
also allows e.g. directly passing in a size that comes from the shape of
an array_t (which is likewise signed).

In order to achieve this, I had to add a templated overload for *all*
integral types and then specialize specifically for `ssize_t` (the
base case) and `size_t` (to remove the -Wnarrowing warning for it,
for convenience/backcompat).  Adding only overloads for `ssize_t`
and `size_t` would not work as `py::bytes{"foo", 3}` would result in
ambiguous overload resolution.