Update deprecated FindPython
CMake 2.27 removes support for deprecated FindPython scripts.

In case CMake is custom built/installed in /usr/local, it seems to pick
up the version of FindPython from the /usr/share folder. This causes
issues when, for example, using a custom built installation of
Python in `/usr/local`: if the version is more recent than what's listed
in the system CMake script, it's not picked up.

Use the new version of the script, so Python is picked up correctly.