Added option to add `return_name` to custom `type_caster`.
This allows to define different python type hints for arguments and
return value. This is implemented using template class `as_return_type`
in order to check if `return_name` is available in the `type_caster` at
compile time. If `return_name` is not available, it falls back to the
previous usage of `name` for both args and return value.