Applied `return_name` to `type_caster` for `std::fileystem::path`.
The `type_caster` previously named `os.PathLike` as both argument and
return type. This is inaccurate since it also takes `str` and `byte` as
argument and actually returns `Path`. This commit uses the new
`return_name` feature to define argument type as
`Union[os.PathLike, str, bytes]` and return type as `Path`.
An assert was added to the unit test for this.