cmake: Use any available Python_add_library command
python_add_library and python3_add_library are mis-cased.
These commands are capitalized.

Even when corrected, this code will fail if Python_FOUND is
set, but Python_add_library() is not available, which appears
to be the issue described here:

 https://github.com/pybind/pybind11/issues/3996

This seems to be due to a mix of FindPython and FindPython3 usage.
Both Python_add_library() and Python3_add_library() support
Python 3.x, so use either one.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>